refactor:Isolate func and model
feat:add confirmation button for batch deletion
